Refactor: move config.inc to bootstrap.php
* Rename this file for clarity (it is used for bootstrapping, not for
  configuration).

* Move this file (and config.php) out of `htdocs/`, since it is not
  intended to be accessed directly on the web.
