+++ phpunit.xml
@@ -2,7 +2,7 @@
 <phpunit
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:noNamespaceSchemaLocation="https://schema.phpunit.de/9.4/phpunit.xsd"
-	bootstrap="src/htdocs/config.inc"
+	bootstrap="src/bootstrap.php"
 	cacheResultFile=".phpunit.cache/test-results"
 	executionOrder="depends,defects"
 	beStrictAboutTodoAnnotatedTests="true"
