+++ src/htdocs/album/album_comment.php
@@ -6,7 +6,7 @@ function create_error_offline($msg) {
 }
 
 try {
-	require_once('../config.inc');
+	require_once('../../bootstrap.php');
 	require_once(LIB . 'Default/smr.inc');
 	require_once(LIB . 'Album/album_functions.php');
