+++ src/htdocs/error.php
@@ -1,5 +1,5 @@
 <?php declare(strict_types=1);
-require_once('config.inc');
+require_once('../bootstrap.php');
 ?>
 <!DOCTYPE html>
