+++ src/htdocs/login.php
@@ -2,7 +2,7 @@
 
 try {
 
-	require_once('config.inc');
+	require_once('../bootstrap.php');
 	require_once(LIB . 'Default/smr.inc');
 
 	// ********************************
