+++ src/htdocs/login_create_processing.php
@@ -1,7 +1,7 @@
 <?php declare(strict_types=1);
 try {
 
-	require_once('config.inc');
+	require_once('../bootstrap.php');
 	require_once(LIB . 'Default/smr.inc');
 
 	if (SmrSession::hasAccount()) {
