+++ src/htdocs/map_warps.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 try {
-	require_once('config.inc');
+	require_once('../bootstrap.php');
 
 	$gameID = Request::getInt('game');
 	if (!SmrSession::hasAccount() || !Globals::isValidGame($gameID)) {
