+++ src/htdocs/resend_password.php
@@ -1,7 +1,7 @@
 <?php declare(strict_types=1);
 
 try {
-	require_once('config.inc');
+	require_once('../bootstrap.php');
 
 	$template = new Template();
 	$template->assign('Body', 'login/resend_password.php');
