+++ src/htdocs/reset_password_processing.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 try {
-	require_once('config.inc');
+	require_once('../bootstrap.php');
 	require_once(LIB . 'Default/smr.inc');
 	
 	$password = Request::get('password');
