+++ src/htdocs/weapon_list.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 try {
-	require_once('config.inc');
+	require_once('../bootstrap.php');
 
 	$template = new Template();
