+++ src/tools/npc/npc.php
@@ -22,7 +22,7 @@ const OVERRIDE_FORWARD = true;
 const NPC_SCRIPT = true;
 
 // global config
-require_once(realpath(dirname(__FILE__)) . '/../../htdocs/config.inc');
+require_once(realpath(dirname(__FILE__)) . '/../../bootstrap.php');
 // bot config
 require_once(CONFIG . 'npc/config.specific.php');
 // needed libs
